python-hello-world-service-6/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Werkzeug==0.16.1
urllib3==1.24.1
PyYAML==5.4.1
python-consul==1.1.0
hvac==0.10.14
psycopg2-binary==2.9.1
msxswagger @ git+https://github.com/CiscoDevNet/python-msx-swagger@v0.6.0
